Portal Gun!
Added a fully functional portal gun. New features still need to be added
but it's in a usable state. To name a few portal position needs to be
adjusted when shot on the corner of a surface, portal needs to be deleted
when shot on surfaces too small, and portal particle needs to ignore
collisions with players/items. Considering a feature where a player shot
with the portal particle is launch through the opposite existing portal
and a feature to make it easier to switch between portal colors.