Initial commit, whole lot to do
Sprite set is a class that holds alll the sprite values for each map. We
will have an array of them in the MainMenu class, which passes it onto
SceneController.
Each map change, it will call void ApplySprites(SpriteSet s), which will
change all sprites in the scene
All floor/wall platforms will be the same. The sprite will be assigned
via ApplySprites() by making an overlap point, seeing what platforms
surround it.
CustomMapObject is simply another class that we use to find platforms,
and contains an identifier saying which kind of platform it belongs to.
I also added a new map called MapTest, which contains all map prefabs.
Not applied yet of course.

TODO:
add it into the main menu method
create method to find out which sprite to use if surrounded by other
platforms
replace all prefabs with default prefab
add in editor script to snap placed items